moved image model reader into appropriate sub folder
